% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regulation_area_function.R
\name{regIndex}
\alias{regIndex}
\title{Calculate fish osmoregulation index according to fitted cubic function}
\usage{
regIndex(kpa, MO2, period, plot = TRUE)
}
\arguments{
\item{kpa}{vector of kpa (x)}

\item{MO2}{vector of MO2 (y)}

\item{period}{vector coded to indicate measurements under "SMR"}

\item{plot}{should a plot of the fitted function be produced?}
}
\value{
list containing fitted model coefficients, Efron's pseudo R2, regulation index, and p50
}
\description{
Calculate fish osmoregulation index according to fitted cubic function
}
\examples{
\dontrun{
# with plot
regIndex(data$kpa, data$MO2, data$period)
# without plot
data.frame(t(unlist(regIndex(data$kpa, data$MO2, data$period, plot = FALSE))))
}
}
